fix: Resolve mypy type errors in stream.py and graph.py
- Remove np.float64() wrapper causing operand type mismatch in stream.py
- Simplify dgl type: ignore to just import-untyped (dgl installed in CI)

+++ src/eventflow/core/adapters/stream.py
@@ -201,7 +201,7 @@ class StreamAdapter(BaseModalityAdapter[StreamOutput]):
 
         # Make timestamps relative to origin
         origin_val: float = float(origin) if origin is not None else 0.0
-        timestamps: np.ndarray = ts_numeric - np.float64(origin_val)
+        timestamps: np.ndarray = ts_numeric - origin_val
 
         # Apply time scaling
         time_scale_params: dict[str, float] = {"origin": origin}
